tried to add flexbox. went poorly
