+++ src/App.css
@@ -35,12 +35,21 @@
   /* color: white; */
 }
 
+.chooseRace {
+  display: flex; /* or inline-flex */
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+  align-content: center;
+  flex-direction: row;
+}
+
 img {
   max-width: 100%;
   max-height: 100%;
 }
 
 .tile {
+  flex-wrap: wrap;
   height: 200px;
   width: 250px;
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
