+++ src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import { Route, HashRouter } from "react-router-dom";
 import eric from "./images/idiot.png";
 import "./App.css";
-import ChooseRace from "./ChooseRace.js";
-import StartPage from "./StartPage.js";
+import ChooseRace from "./Pages/ChooseRace.js";
+import StartPage from "./Pages/StartPage.js";
 import Amplify, { Analytics } from "aws-amplify";
 import aws_exports from "./aws-exports";
 // import { withAuthenticator } from "aws-amplify-react";
