+++ src/Pages/PageNotFoundComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import "./App.css";
+import "../App.css";
 import { Analytics } from "aws-amplify";
 
 class PageNotFoundComponent extends Component {
